% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublefunction.R
\name{doublefunction}
\alias{doublefunction}
\title{Doubles a number}
\usage{
doublefunction(x)
}
\arguments{
\item{x}{A number.}
}
\value{
A number.
}
\description{
Doubles a number
}
\examples{
doublefunction(-10)
}
